refactor: 🔮 Refactored app.photo() and app._process_streaming() to maintain image aspect ratio during resizing
The code changes include making sure that the aspect ratio of images is maintained when resizing in both the app.photo() and app._process_streaming() functions. A new parameter `max_width` and `max_height` were also included in the app.photo() function to ensure that the image does not exceed those dimensions when resizing. The maintain_aspect_ratio_resize() function in sleepy_baby/helpers.py was also refactored to accept width and height parameters and select the smallest format.